==> analyses/3.1. Robustness_overlap.rmd <==


```{r}
library(tidyverse)
options(digitis = 3)
options(scipen = 10^5)
options(help_type = "html")
library(tidyverse)
source("R/generalB.r")
source("R/princeB.r")
library(survey)
library(dbarts)
library(haven)
library(parallel)
```

Stacked source and dhs data
```{r}
#dhs data
dhs <- readRDS("data/dhs_plus_mock.rds") %>%
  filter(anymodern_ever == 0 & anymodern_now == 0 & want_no_birth == 1)

#source study data
sdt <- readRDS("data/from_source_study/df.rds") %>%
  rename(wt = complier_wt)

#propensity to be in 4 cities from source
me <- readRDS("data/from_source_study/propensity_bart.rds")
sdt$e <- apply(predict(me, newdata = sdt, type = "bart"), 2, mean)
sdt$urban <- sdt$in5states <- NA


common_cols <- intersect(colnames(dhs), colnames(sdt))
stacked_dt <-
  bind_rows(
    dhs %>% select(all_of(common_cols))
    , sdt %>% select(all_of(common_cols))
    , .id = "source"
  ) %>%
  mutate(source = as.numeric(source) - 1)

table(stacked_dt$source)

```


estimate propensity to be in source P(T=0|X, T=1 or T=0)
```{r}

set.seed(0203)
m_pi <- bart2(source ~ .
  , data = select(stacked_dt, - c(urban, wt, in5states))
)
sort(apply(m_pi$varcount, 3, mean), decreasing = TRUE)

pi_t <- fitted(m_pi)
tapply(pi_t, stacked_dt$source, mean)
tapply(pi_t, stacked_dt$source, sd)


#stacked_dt$pi_t <- pi_t
#with(stacked_dt, tapply(work_last_year,source,mean))
```


compute complier propensity P(C=1|T=0)
(Note that this is the probability of being a complier as a funciton of X
not the posterior pedictive probability of being a complier given all obsered data, including Y)
```{r}
pi_c <- lapply(1:4, \(i) {
  stree <- readRDS(paste0("data/from_source_study/tree_", i, ".rds"))

  scaled_x <- stacked_dt %>%
    select(all_of(names(attr(stree, "scaled:center")))) %>%
    scale(attr(stree, "scaled:center"), attr(stree, "scaled:scale"))

  eco <- predict(subset(stree, m == "co"), scaled_x, n_cores = 20)
  eco
})

stacked_dt$pi_c <- apply(pi_c %>% bind_cols, 1, mean)
tapply(stacked_dt$pi_c, stacked_dt$source, mean)
```



```{r}
library(scales)
stacked_dt$pi_s <- with(stacked_dt, pi_c * pi_t)
tapply(stacked_dt$pi_s, stacked_dt$source, mean)
stacked_dt$e_s <- qlogis(stacked_dt$pi_s)

m <- with(filter(stacked_dt, source == 1), weighted.mean(e_s, wt))
m2 <- with(filter(stacked_dt, source == 1), weighted.mean((e_s)^2, wt))
s <- sqrt(m2 - m^2)
stacked_dt$e_s_tilde <- (stacked_dt$e_s - m) / s


oplot <- ggplot() +
  geom_density(
    data = data.frame(
      filter(stacked_dt, source == 0, urban == 1, in5states == 1)
      , pop = "2"
    )
    , aes(x = e_s_tilde, weight = wt, fill = pop)
    , alpha = .4, color = NA
  ) +
  geom_density(
    data = data.frame(filter(stacked_dt, source == 0), pop = "4")
    , aes(x = e_s_tilde, weight = wt, fill = pop)
    , alpha = .4, color = NA
  ) +
  geom_density(
    data = data.frame(filter(stacked_dt, source == 0, urban == 1), pop = "3")
    , aes(x = e_s_tilde, weight = wt, fill = pop)
    , alpha = .4, color = NA
  ) +
  geom_density(
    data = data.frame(filter(stacked_dt, source == 1), pop = "1")
    , aes(x = e_s_tilde, weight = wt, fill = pop)
    , alpha = .4, color = NA
  ) +
  xlim(-10, 5) +
  xlab("standardized selection score") +
  ylab("density") +
  theme_minimal() +
  theme(legend.position = "inside", legend.position.inside = c(.3, .8)) +
  theme(legend.title = element_blank()) +
  scale_fill_manual(values = hue_pal()(4)
    , name = element_blank()
    , labels = c(
    "Compliers (source study)"
    , "Urban, same states (DHS)"
    , "Urban, all state (DHS)"
    , "Urban and rural, all states (DHS)"
    )
  )

oplot
#png("output/overlap.png"
#    , width = 480 * 3, heigh = 480 * 2.5, res = 300)
#oplot
#dev.off()
```
