==> analyses/3.2. Robustness_confounding.rmd <==
```{r}
library(tidyverse)
options(digitis = 3)
options(scipen = 10^5)
options(help_type = "html")
options(mc.cores = parallel::detectCores())

library(haven)
library(parallel)
library(gt)
library(tidyverse)
source("R/generalB.r")
source("R/princeB.r")
```

URBAN
```{r}
dhs <- readRDS("data/dhs_plus_mock.rds")
colnames(dhs)

dhs$sp <- dhs$anymodern_ever == 0 &
  dhs$anymodern_now == 0 &
  dhs$want_no_birth == 1 &
  dhs$urban == 1
```

====================================
***Simple Sensitivity Analysis***
====================================
```{r}
set.seed(0203)
pate_sa <- sapply(seq(0, 1, by = .1)
  , \(prevalence) gen_simple_sim(with(dhs, y1 - y0), prev = prevalence)
)
pate_sa <- as.data.frame(t(pate_sa))
pate_sa$p <- seq(0, 1, by = .1)
pate_sa
```


```{r}
conf_plot <-
  pate_sa %>%
  ggplot(aes(y = m, x = p)) +
  geom_hline(yintercept = 0, col = "grey") +
  geom_errorbar(aes(ymin = l, ymax = u),  width = .05) +
  geom_point() +
  labs(title = ""
    , x = "Prevalence of confounding variable (ξ)"
    , y = "PATE"
  ) +
  theme_minimal() +
  theme(legend.position = "none")

conf_plot

#png("output/conf_plot.png"
#  , width = 480 * 5, heigh = 480 * 2.5, res = 300
#)
#conf_plot
#dev.off()
```


=====================================
***Nonparametric Sensitivity Analysis***
=====================================
```{r}
set.seed(0203)
s <- sample(ncol(dhs$y1), 100) #small sample from posterior
g <- exp(seq(0, 3, by = .5))
lw <- sapply(g, \(x) gen_shift_sim(with(dhs, y1 - y0)[, s], γ = x, inf = TRUE))
up <- sapply(g, \(x) gen_shift_sim(with(dhs, y1 - y0)[, s], γ = x, inf = FALSE))

#fortmatting
lw_sa <- data.frame(t(lw))
up_sa <- data.frame(t(up))
pate_sa <- bind_cols(list(lw_sa, up_sa))
lbs <- expand.grid(c("m", "s", "l", "u"), c("lw", "up"))
colnames(pate_sa) <- apply(lbs, 1, paste0, collapse = "_")
pate_sa$g <- log(g)
pate_sa$m0 <- pate_sa$m_lw[1]
pate_sa
#saveRDS(pate_sa, "output/pate_sa_gamma.rds")
```

```{r}
#pate_sa <- readRDS("output/pate_sa_gamma.rds")

conf_plot <-
  pate_sa %>%
  ggplot(aes(y = m0, x = g)) +
  geom_hline(yintercept = 0, col = "grey") +
  geom_errorbar(aes(ymin = m_lw, ymax = m_up),  width = .025, size = 1) +
  geom_errorbar(aes(ymin = l_lw, ymax = u_up),  width = .05
    , alpha = .6
  ) +
  labs(title = ""
    , x = "log(Γ)"
    , y = "PATE"
  ) +
  theme_minimal() +
  theme(legend.position = "none")

conf_plot

#png("output/conf_plot_gamma.png"
#  , width = 480 * 5, heigh = 480 * 2.5, res = 300
#)
#conf_plot
#dev.off()
```
